% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsdl.R
\name{subset.tsdl}
\alias{subset.tsdl}
\title{Subset of time series from the Time Series Data Library}
\usage{
\method{subset}{tsdl}(x, cond1, cond2, ...)
}
\arguments{
\item{x}{TSDL data or a subset of TSDL data}

\item{cond1}{Subject (type) or frequency of the data. Subject is a character variable and
frequency is a numeric variable.}

\item{cond2}{Optional second condition specifying type or period of the
data, depending on \code{cond1}.  If \code{cond1} denotes subject then
\code{cond2} would denote frequency, but if \code{cond1} denotes frequency then
\code{cond2} would denote subject.}

\item{...}{Other arguments.}
}
\value{
An object of class \code{tsdl} consisting of the selected series.
}
\description{
\code{subset.tsdl} returns a subset of the time series data from
the Time Series Data Library. Subsets can be for specific periods,
or specific types of data or both.
}
\details{
See \code{unique(meta_tsdl$frequency)} for possible values
for \code{cond1} and \code{cond2} denoting frequency.

See \code{unique(meta_tsdl$subject)} for possible values
for \code{cond1} and \code{cond2} denoting subject.

Partial matching used for both conditions.
}
\examples{

tsdl.quarterly <- subset(tsdl,4)
tsdl.daily.industry <- subset(tsdl,365,"Industry")
}
\seealso{
\code{\link{tsdl}}

\code{\link{meta_tsdl}}
}
\keyword{data}
