% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsdl-package.R
\docType{package}
\name{tsdl-package}
\alias{tsdl-package}
\title{Time Series Data Library}
\source{
\url{https://datamarket.com/data/list/?q=provider:tsdl}.
}
\description{
The 648 time series from the Time Series Data Library.
The Time Series Data Library (TSDL) was created by Rob Hyndman,
Professor of Statistics at Monash University, Australia.
}
\references{
Hyndman, R.J. "Time Series Data Library",
\url{https://datamarket.com/data/list/?q=provider:tsdl}.
}
\author{
Rob J Hyndman. \email{Rob.Hyndman@monash.edu}

Yangzhuoran Yang. \email{Fin.Yang@monash.edu}
}
\keyword{package}
