% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.tsdl}
\alias{subset.tsdl}
\title{Subset of time series from the Time Series Data Library}
\usage{
\method{subset}{tsdl}(x, cond1, cond2, ...)
}
\arguments{
\item{x}{TSDL data or a subset of TSDL data}

\item{cond1, cond2}{Optional conditions specifying subject (type) or frequency of the data.
Character variable would be recognised as subject
and numeric variable would be recognised as frequency. Positions are interchangeable.}

\item{...}{Other arguments specifying conditions for where to search and what to search.
\describe{
\item{\code{start}}{Numeric variable specifying starting year of the series.}
\item{\code{description}}{String to be matched in the description attribute of the time series.}
\item{\code{source}}{String to be matched in the source attribute of the time series.}
}}
}
\value{
An object of class \code{tsdl} consisting of the selected series.
}
\description{
\code{subset.tsdl} returns a subset of the time series data from
the Time Series Data Library. Subsets can be for specific periods,
or specific types of data or both.
}
\details{
See \code{unique(meta_tsdl$frequency)} for possible values
for \code{cond1} and \code{cond2} denoting frequency.

See \code{unique(meta_tsdl$subject)} for possible values
for \code{cond1} and \code{cond2} denoting subject.

Partial matching used for both conditions.
}
\examples{
# Subset by frequency
tsdl_quarterly <- subset(tsdl,4)
tsdl_quarterly

# Subset by frequency and subject
tsdl_daily_industry <- subset(tsdl,12,"Industry")
tsdl_daily_industry

# Subset by source
tsdl_abs <- subset(tsdl, source = "Australian Bureau of Statistics")
tsdl_abs

# Subset by starting year
tsdl_1948 <- subset(tsdl, start = 1948)
tsdl_1948

# Subset by description
tsdl_nettraffic <- subset(tsdl, description = "Internet traffic")
tsdl_nettraffic

}
\seealso{
\code{\link{tsdl}}

\code{\link{meta_tsdl}}
}
\author{
Yangzhuoran Yang
}
\keyword{data}
